Updated the delete user one because transaction book is directly related to user ,so delete the entry in transaction book first and then delete user(now deleting transaction book record for only specific user)

+++ src/main/java/com/LibraryManagementSystem/LMS/project/Controller/UserController.java
@@ -158,7 +158,9 @@ public class UserController {
 
         List<transaction_book> transactionBooks = tbService.getAllTransaction_Book();
         for (transaction_book transactionBook : transactionBooks) {
-            tbService.deleteTransactionBook(transactionBook.getId());
+            if (transactionBook.getTransaction_id().getcard_id().getCustomer().getUser().getId() == id) {
+                tbService.deleteTransactionBook(transactionBook.getId());
+            }
         }
 
         userService.deleteUser(id);
